# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Translation Management System APIs",
        "description": "Your API Description",
        "contact": {
            "name": "API Support",
            "email": "nahyan.bin.khalid@gmail.com"
        },
        "license": {
            "name": "Nahyan",
            "url": "nahyan.bin.khlid@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/tms/public"
        }
    ],
    "paths": {
...
        "/api/v1/translations": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "List of Translations",
                "operationId": "c0961b0a88aa0196c0baab548f6669a1",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Locale code",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "en"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "Tag",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "mobile"
                        }
                    },
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Translation key",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "home.title"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "welcome"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile Details"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Add a Translation",
                "operationId": "67bc06bb460f643e572547074c9003dc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "locale",
                                    "key",
                                    "content",
                                    "tags"
                                ],
                                "properties": {
                                    "locale": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "home.title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Welcome"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "mobile",
                                            "web"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation Added Successfully"
                    },
                    "401": {
                        "description": "Data Incomplete or Invalid"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/translations/{id}": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Show Translation",
                "operationId": "1e5f551a06bb3009e67431c5c542843a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile Details"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update a Translation",
                "operationId": "0ae3a6a5011658925866595838434c55",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "locale",
                                    "key",
                                    "content",
                                    "tags"
                                ],
                                "properties": {
                                    "locale": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "key": {
                                        "type": "string",
                                        "example": "home.title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Welcome"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "mobile",
                                            "web"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Registered Successfully"
                    },
                    "401": {
                        "description": "Data Incomplete or Invalid"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Translations"
                ],
                "summary": "Delete a Translation",
                "operationId": "a0deb6a75764e8d34cb9cdf87184aa1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Registered Successfully"
                    },
                    "401": {
                        "description": "Data Incomplete or Invalid"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
...
}